Split out spark job service tests by method
